fix up the `collaborate` page
